// ClientApp/src/app/models/TvShow.ts
export interface TvShow {
    id: number;
    backdrop_path?: string;
    firstAirDate?: Date;
    name?: string;
    origin_country?: string;
    original_language?: string;
    original_name?: string;
    overview?: string;
    popularity: number;
    poster_path?: string;
    vote_average: number;
    vote_count: number;
}